experiments/image_bias/requirement_tunning.txt
# core
torch==2.2.2
torchvision==0.17.2
torchaudio==2.2.2

# diffusion + HF stack
diffusers==0.34.0
transformers==4.55.2
accelerate==0.30.1
peft==0.12.0
huggingface-hub>=0.23,<0.35
safetensors>=0.4,<0.7
tokenizers>=0.20,<0.22
sentencepiece>=0.2.0

# common pins to avoid resolver fights
protobuf<5
numpy==1.26.4
pillow>=10,<12
tqdm>=4.66,<5
requests>=2.31,<3
python-dotenv>=1.0,<2
